MODFEE-194: DaoImpl: Use single id methods for get, delete, update
Running a PostgreSQL get, delete or update query with a single id
should use the primary key id column of the table. The primary key
automatically has a search index.

The DaoImpl uses the id property of the JSON in the jsonb field that doesn't have a search index.

PostgresClient provides dedicated get, delete and update methods that are faster and easier to use.